Services cambios

+++ src/main/java/pe/idat/RetoGit/service/AlumnosService.java
@@ -1,11 +1,15 @@
 package pe.idat.RetoGit.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import pe.idat.RetoGit.model.Alumnos;
 import pe.idat.RetoGit.repository.AlumnosRepository;
 
 import java.util.ArrayList;
+import java.util.Optional;
 
 @Service
 public class AlumnosService {
@@ -28,11 +32,36 @@ public class AlumnosService {
     }
 
     public boolean eliminarAlumno(Long id){
-        try {
+        if (alumnosRepository.existsById(id)) {
             alumnosRepository.deleteById(id);
             return true;
-        } catch (Exception err){
+        } else {
             return false;
+        }
+    }
+
+    public Alumnos modificarAlumno(Long id, Alumnos alumnoActualizado){
+        Optional<Alumnos> alumnoExistente = alumnosRepository.findById(id);
+
+        if (alumnoExistente.isPresent()) {
+            Alumnos alumno = alumnoExistente.get();
+
+            if (alumnoActualizado.getNombre() != null){
+                alumno.setNombre(alumnoActualizado.getNombre());
+            }
+            if (alumnoActualizado.getApellidos() != null){
+                alumno.setApellidos(alumnoActualizado.getApellidos());
+            }
+            if (alumnoActualizado.getFechaNac() != null){
+                alumno.setFechaNac(alumnoActualizado.getFechaNac());
+            }
+            if (alumnoActualizado.getColegio() != null){
+                alumno.setColegio(alumnoActualizado.getColegio());
+            }
+
+            return alumnosRepository.save(alumno);
+        } else {
+            return null;
         }
     }
 }
